requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.8 # Apache-2.0
cliff>=2.3.0
shade>=1.12.0
os-client-config>=1.22.0
PyYAML>=3.10.0
progressbar2>=3.12.0
